requirements.txt
numpy~=1.26.2
matplotlib~=3.8.2
scikit-learn~=1.3.2
imageio~=2.33.0
colormath~=3.0.0
mediapipe~=0.10.8
opencv-python~=4.8.1.78
Django~=4.2.7
djangorestframework
requests